+++ include/sym.h
@@ -30,7 +30,7 @@ extern long 	sym_destroy(char *key);
 extern int 	sym_insert(char *key, int symtype, long symdata);
 extern SYMBOL 	sym_find(char *key, char **key_ret);
 extern int 	hash_create(unsigned nel);
-extern int 	MakeInternalSym(char *name, char ed_typ, long value);
+extern int 	MakeInternalSym(char *name, char ed_typ, intptr_t value);
 
 #define	SYM_H
 # endif
